Fix use of OS-dependent File.separatorChar in FileUtils
The paths given to this method may use a separator other than the one
present in File.separatorChar at runtime. Particularly, the paths in the
legacy classpath is always written out with '/' (forward slash), meaning
the use of File.separatorChar causes this method to fail when on
Windows, where the value is `\` (backslash).

Instead of using the File.separatorChar or regex-ing on both known
separators, we delegate instead to the JDK's own Path interface.