Compute jar module descriptors in parallel (#57)
To compute a module descriptor, we need to compute the package list, and
that is often slow. So do it in parallel! (And defer computation of the
module descriptor until it is actually needed).

Also added a new method to retrieve the provided services without
computing the full module descriptor, for use in the
`ModDirTransformerDiscoverer`.

Tested in ATM8.