Test `ModInfo` i18n behavior
We cannot depend on `I18nExtension` here, so the test mocks `I18NParser` and assumes that a parser will return the i18n-key when a translation does not exist.

Since it is a unit test, not an integration test, this is acceptable IMO.

The `Bindings` singleton was updated so that it's static methods can be mocked without an instance being created.