Reject invalid service files (#62)
Currently, SJH throws when building a module descriptor for jars which
contain files in META-INF/services which aren't valid. All this PR does
is filter out these invalid files:
- I've added a set of known "naughty" service files which don't follow
the format correctly - It was mentioned in Discord that certain Groovy
libraries define extensions with these invalid service files so I wanted
to include these too
- JLine 3.22.0 and above define a nested file in
`META-INF/services/org/jline/terminal/provider/` which doesn't follow
the service file format (it's in fact a properties file). This causes
SJH to throw because `exec` is not within a named package, and is
blocking upgrading to newer versions of JLine