Always load UnionURLStreamHandler even in non-modular environments (#74)
The use case for this is the use of SJH inside of FML unit tests where
SJH itself was not loaded as a module.

It is perfectly capable of being used, but fails due to the `union`
protocol handler not being loaded. This change makes SJH always load its
own protocol handler regardless of layer and supports the extension
point in non modular environments too.

As an aside: The extension point does not seem to be used by anyone else
according to a GH search and can likely be deprecated/removed at some
point.