Increase log verbosity of refused event dispatch messages
Usually, when a modder has an error in their code, we will spam out a bunch of these errors because the mod state is broken.

However, if it were a true error, you would expect all of our dispatches to be guarded with something to the effect of `if (!bus.isInErrorState()) { bus.dispatchEvent(...); }` which is not the case.

Because of this, it's noise at best and should be warn at highest. Instead, I chose trace because it's TYPICALLY only of use to Neo devs, or more involved debugging from modders.